rename w3sl_noevm to wsloop
