+++ assets/agents/test/test_MinterAgents.py
@@ -1,7 +1,7 @@
 from enforce_typing import enforce_types
 
 from assets.agents.MinterAgents import *
-from assets.netlists.w3sl_noevm import SimState, SimStrategy
+from assets.netlists.wsloop import SimState, SimStrategy
 from engine import AgentBase
 from util.constants import BITCOIN_NUM_HALF_LIVES, \
     S_PER_DAY, S_PER_MONTH, S_PER_YEAR
