+++ assets/netlists/wsloop.py
@@ -1,6 +1,5 @@
 """
-w3sl_noevm:
-Netlist to simulate Web3 Sustainability Loop (w3sl), with no EVM
+Netlist to simulate Web3 Sustainability Loop, with no EVM
 
 Implements:
 -class SimStrategy(SimStrategyBase)
